+++ app/api/utils/__init__.py
@@ -3,15 +3,7 @@
 # This program is distributed under the terms of the GNU General Public License: GPL-3.0-or-later  #
 # ------------------------------------------------------------------------------------------------ #
 
-"""URL configuration for the Django project."""
+"""API Utilities package.
 
-from django.conf import settings
-from django.conf.urls.static import static
-from django.urls import include, path
-from django.views.generic import RedirectView
-
-urlpatterns = [
-    path('', RedirectView.as_view(url='/api/', permanent=True)),
-    path('api/', include('api.urls')),
-    *static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT),
-]
+This package provides utility and helper functions for the pseudonymizing API.
+"""